docker: adds script to start docker with volume
dockerdev.sh starts the docker container based on shift/mcu-base with thir
directory as a volume. This commit also removed the copy workspace and build
commands from the Dockerfile.dev, since this is not needed any longer. When
using volumes, one can easily access files produced in the docker container and
the docker container can see updates in the volumes, which means that the image
doesn't have to be re-built to produce a firmware binary.