communication: added HMAC on encrypted messages to and from the BitBox
Appending an HMAC on the encrypted messages to and from the BitBox.
This should mitigate AES malleability on the first AES block since
the encrypted content will now be checked for manipulation.
Key derivation for the encryption and authentication key is performed on the
shared secret derived from the device password.