commander: separate main and hidden wallet backup checks
Before, checking the backup recovery password when logged in using the
main password would succeed for both the main and the hidden wallet
recovery password. This was the only way to check the hidden wallet
password, as everything sd-card was in the hidden wallet.

With this commit, one can only check the main recovery password in the
main wallet, and the hidden recovery password when in the hidden
wallet.